requirements.txt
absl-py==1.0.0
affine==2.3.1
aiosignal==1.2.0
astunparse==1.6.3
attrs==21.4.0
cachetools==5.0.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
cycler==0.11.0
distlib==0.3.4
docker-pycreds==0.4.0
filelock==3.7.0
flatbuffers==2.0
fonttools==4.33.3
frozenlist==1.3.0
gast==0.5.3
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
h5py==3.6.0
idna==3.3
image-similarity-measures==0.3.5
imageio==2.18.0
importlib-metadata==4.11.3
importlib-resources==5.7.1
javabridge==1.0.19
jsonschema==4.5.1
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
libclang==14.0.1
Markdown==3.3.6
matplotlib==3.5.1
msgpack==1.0.3
networkx==2.8
numpy==1.22.3
oauthlib==3.2.0
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
patchify==0.2.3
pathtools==0.1.2
phasepack==1.5
Pillow==9.1.0
platformdirs==2.5.2
promise==2.3
protobuf==3.20.1
psutil==5.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyFFTW==0.13.0
pyparsing==3.0.8
pyrsistent==0.18.1
python-dateutil==2.8.2
PyWavelets==1.3.0
PyYAML==6.0
rasterio==1.3.3
ray==1.12.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.19.2
scipy==1.8.0
sentry-sdk==1.5.12
setproctitle==1.2.3
shortuuid==1.0.9
six==1.16.0
smmap==5.0.0
snuggs==1.4.7
split-folders==0.5.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.25.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
tifffile==2022.4.28
tqdm==4.64.0
typing_extensions==4.2.0
urllib3==1.26.9
virtualenv==20.14.1
wandb==0.12.17
Werkzeug==2.1.2
wrapt==1.14.0
zipp==3.8.0
